pyproject.toml
[tool.poetry]
name = "gh"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0.0"
lxml = "^5.1.0"
scrapy = "^2.11.0"
requests = "^2.31.0"
pandas = "^2.2.0"
googlemaps = "^4.10.0"
datetime = "^5.4"
dash = "^2.16.0"
plotly-express = "^0.4.1"
black = "^24.2.0"
matplotlib = "^3.8.3"
seaborn = "^0.13.2"
urllib3 = "^2.2.1"
jsons = "^1.6.3"




[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
